+++ OpenTibia.Server/BodyContainerItem.cs
@@ -196,7 +196,7 @@ namespace OpenTibia.Server
 
                 // Valid target, special slots
                 case Slot.LeftHand:
-                    if (!(player.Inventory[Slot.RightHand] is IContainerItem rightHandContainer))
+                    if (!(player.Inventory[(byte)Slot.RightHand] is IContainerItem rightHandContainer))
                     {
                         return false;
                     }
@@ -206,7 +206,7 @@ namespace OpenTibia.Server
                     return rightHandItem == null || (item.DressPosition != Slot.TwoHanded && rightHandItem.DressPosition != Slot.TwoHanded);
 
                 case Slot.RightHand:
-                    if (!(player.Inventory[Slot.LeftHand] is IContainerItem leftHandContainer))
+                    if (!(player.Inventory[(byte)Slot.LeftHand] is IContainerItem leftHandContainer))
                     {
                         return false;
                     }
