+++ OpenTibia.Server/MovementEvents/BodyToBodyMovementEvent.cs
@@ -68,7 +68,7 @@ namespace OpenTibia.Server.MovementEvents
             {
                 bool moveSuccessful = thingMoving is IItem item &&
                                       creatureFinder.FindCreatureById(targetCreatureId) is IPlayer targetPlayer &&
-                                      this.Game.PerformItemMovement(item, targetPlayer.Inventory[fromCreatureSlot] as IContainerItem, targetPlayer.Inventory[toCreatureSlot] as IContainerItem, 0, 0, amount, this.Requestor);
+                                      this.Game.PerformItemMovement(item, targetPlayer.Inventory[(byte)fromCreatureSlot] as IContainerItem, targetPlayer.Inventory[(byte)toCreatureSlot] as IContainerItem, 0, 0, amount, this.Requestor);
 
                 if (!moveSuccessful)
                 {
