+++ OpenTibia.Server/MovementEvents/EventConditions/CanDressThingAtTargetSlotEventCondition.cs
@@ -88,7 +88,7 @@ namespace OpenTibia.Server.MovementEvents.EventConditions
 
                 // Valid target, special slots
                 case Slot.LeftHand:
-                    if (!(player.Inventory[Slot.RightHand] is IContainerItem rightHandContainer))
+                    if (!(player.Inventory[(byte)Slot.RightHand] is IContainerItem rightHandContainer))
                     {
                         return false;
                     }
@@ -98,7 +98,7 @@ namespace OpenTibia.Server.MovementEvents.EventConditions
                     return rightHandItem == null || (item.DressPosition != Slot.TwoHanded && rightHandItem.DressPosition != Slot.TwoHanded);
 
                 case Slot.RightHand:
-                    if (!(player.Inventory[Slot.LeftHand] is IContainerItem leftHandContainer))
+                    if (!(player.Inventory[(byte)Slot.LeftHand] is IContainerItem leftHandContainer))
                     {
                         return false;
                     }
